feat: 트랜잭션 서비스 추상화 - 트랜잭션 로직과 서비스 로직 분리
