+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,69 +1,10 @@
 package com.techcourse.service;
 
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.interface21.transaction.support.TransactionSynchronizationManager;
-import com.techcourse.config.DataSourceConfig;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
+public interface UserService {
 
-public class UserService {
-
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
-    private static final DataSource dataSource = DataSourceConfig.getInstance();
-
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        try {
-            connection.setAutoCommit(false);
-            try {
-                final var user = findById(id);
-                user.changePassword(newPassword);
-                userDao.update(user);
-                userHistoryDao.log(new UserHistory(user, createBy));
-            } catch (Exception e) {
-                rollback(e, connection);
-                throw e;
-            }
-            connection.commit();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e.getMessage(), e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-            TransactionSynchronizationManager.unbindResource(dataSource);
-        }
-    }
-
-    private void rollback(Exception appException, Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException rollbackEx) {
-            log.error(rollbackEx.getMessage(), rollbackEx);
-            appException.addSuppressed(rollbackEx);
-        }
-    }
+    User findById(final long id);
+    void save(final User user);
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
